fix(core): fix `ctx.url` have extra trailing question mark when path and querystring is emtpy
This commit addresses the issue of ctx.url returning incorrect format when the path is empty in koam-core module. We have added a unit test to ensure correct url format is returned even when the path is empty. Further modifications include syntax enhancements in index.ts and request.ts. The MiddlewareGenerator type now accepts any number of any type of arguments, enhancing its flexibility. Also rectified the url getter in request.ts to not append a trailing '?' when querystring is empty.